moved string_to_bool to tools_paths
