+++ demisto_sdk/commands/common/tests/tools_test.py
@@ -93,7 +93,6 @@ from demisto_sdk.commands.common.tools import (
     search_and_delete_from_conf,
     server_version_compare,
     str2bool,
-    string_to_bool,
     to_kebab_case,
 )
 from demisto_sdk.tests.constants_test import (
@@ -2529,27 +2528,6 @@ def test_get_display_name(data, answer, tmpdir):
     """
     file = File(tmpdir / "test_file.json", "", json.dumps(data))
     assert get_display_name(file.path) == answer
-
-
-@pytest.mark.parametrize("value", ("true", "True", 1, "1", "yes", "y"))
-def test_string_to_bool_true(value: str):
-    assert string_to_bool(value)
-
-
-@pytest.mark.parametrize("value", ("", None))
-def test_string_to_bool_default_true(value: str):
-    assert string_to_bool(value, True)
-
-
-@pytest.mark.parametrize("value", ("false", "False", 0, "0", "n", "no"))
-def test_string_to_bool_false(value: str):
-    assert not string_to_bool(value)
-
-
-@pytest.mark.parametrize("value", ("", " ", "כן", None, "None"))
-def test_string_to_bool_error(value: str):
-    with pytest.raises(ValueError):
-        string_to_bool(value)
 
 
 @pytest.mark.parametrize(
